moves EntryPoint to submodule and adds a test
